libcamera2: restrict preview update mode to standard in online captures
BZ: 127752

Problem:
In the following scenario:
- user requests a preview update mode that requires continuous capture mode
- other conflicting setting force the hal to drop to old online mode

The preview was not being updated.
This is what it was happening with SocialCamera that was
requesting continuous update mode in night scene mode.

Night mode requires ANR that is not supported in continuous mode.

Solution:
All preview update modes can be supported in continuous preview mode
but only the standard update mode is supported if we are in traditional
online preview mode

This patch enforces this limitation so that preview is updataed in this case

Change-Id: If1e9b5b5a728fe03f57ba4a0b1426c06e213b8f0
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/128784
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>