libcamera2: Use graphic recording buffer for Gen
BZ: 102512

Video encoder needs NV12 tiled format in BYT but ISP outputs NV12
linear only. So we need to do color convertion here.
We use CPU to convert previously and now to improve the
performance,GPU can be used to do this work. GPU need to operate graphic
buffer, so this patch is to allocate graphic memory for recording buffer
and do color convertion before post data to encoder.
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>

Change-Id: I7dba75c6bbcec829e2c07a663ced2db082ce8d66
Reviewed-on: http://android.intel.com:8080/132509
Reviewed-by: Xie, Hongbiao <hongbiao.xie@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>