libcamera2: Burst capture memory optimizing
BZ: 138433

Camera burst capture is the most typical case of memory consumption
because it will allocate 10 YUV buffers in worst case. These buffers are
only used once.

Usually, The number of snapshot buffers is not necessary equal to burst
length.Make this value configurable by camera_profiles.xml. Set the
default value to 10 to make sure it doesn't affect the product which
has not been tested fully.

Snapshot buffer from now on would be retruned to ISP if necessary to
make it to be used again.

Change-Id: I7cdb4ea0b8d8ec051dad2abca693f63a7fd3ef83
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>
Reviewed-on: http://android.intel.com:8080/132541
Reviewed-by: Yang, Liang L <liang.l.yang@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>