libcamera2: [REVERTME] drop corrupt AE statistics
BZ: 139972

Implements a heuristic algorithm for dropping ISP statistics,
depending on the variation between previous and current statistics.
The threshold for dropping the statistics is adjustable, based on
perceived performance. Also the statistics should not be dropped
during the (pre-)flash sequence.

This patch cleans up the previous implementation, as well as
fixes a regression in the image exposure when using flash.
The regression was because of an extra 3A iteration in the pre-flash
sequence due to additional applyPreFlashProcess() call.

NOTE: Revert once BZ 119181 is resolved.

Based on change I1c7d354bb0adbb352d66d0b67a9c7b80d5f4ee81

Change-Id: I34244f863d94575fdcadf0ff7e6966c19211d710
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Reviewed-on: http://android.intel.com:8080/133377
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Mandy, Arnaud <arnaud.mandy@intel.com>
Reviewed-by: Patale, Pavan <pavan.patale@intel.com>
Tested-by: Patale, Pavan <pavan.patale@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>