libcamera2: Use graphic recording buffer
BZ: 102512

This patch is going to change the recording buffer allocation place from
MemoryHeap to Graphic. Graphic memory is convenience for both software
and hardware encoder to access. It keeps the old way that share virtual
address to IntelMetaDatabuffer and would be beneficial to the encoder
that want to use Graphic buffer directly in future. Non-cacheable
gralloc buffers also would save power.

Change-Id: I1482b867ffe4140771e21ad7692bda7978c50dad
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>
Reviewed-on: http://android.intel.com:8080/133910
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: Yang, Liang L <liang.l.yang@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>