libcamera2: Provide NV12 tiled recording buffer for Gen
BZ: 102512

To optimaze video recording, video encoder in Baytrail(with Gen GPU)
needs NV12 tiled format source data but ISP supports NV12 linear
only.The unavoidable color conversion was implemented by CPU previously
and now to improve the performance, GPU can be used to do this work.
Note that:
NV12 tiled is a special format for video encoding. Camera is not
recommanded to offer this kind of data and even do conversion as well.
Hope encoder can do the private work in the fulture OR our ISP is able to
output tiled data directly.

Change-Id: Ia6e8e4d83a1587c6a0da63c7ecc7b4413c64ad21
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>
Reviewed-on: http://android.intel.com:8080/133912
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>