libcamera2: Provide NV12 tiled recording buffer for Gen
BZ: 102512

To optimize video recording, video encoder in Baytrail(with Gen GPU)
needs NV12 tiled format source data but ISP supports NV12 linear
only.The unavoidable color conversion was implemented by CPU previously
and now to improve the performance, GPU can be used to do this work.
NOTE: NV12 tiled is a special format for video encoding. Camera is not
recommended to offer this kind of data and even do conversion as well.
Hope encoder can do the private work in the future OR our ISP is able to
output tiled data directly.

Change-Id: I1a434694b17b88e5ca6d1964a77a77b7c6b787e6
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/137204
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>