libcamera2: Introduce AtomBuffer as format descriptor
BZ: 142982

AtomBuffer is already a structure describing the format of buffer.
AtomBuffer is also used as a container for delivering frames inside
Camera HAL. Components responsible of handling AtomBuffer are anyhow
configured with different ways. This makes pre-planned refactoring
and higher-level controlling of pipelines difficult.

This change introduces
AtomBufferType::ATOM_BUFFER_FORMAT_DESCRIPTOR = 0

Clear AtomBuffer structure is hereby automatically a format
descriptor. When memory is associated to it, the type must be set
accordingly.

Initially FrameInfo structure is dropped and replaced with
AtomBuffers as format descriptors. Further work is expected to
utilize AtomBuffers rather than distinct variables or overlapping
structures.

Change-Id: I62df64308aca7b33fddc79c8bb1388a42cbe0e5f
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Reviewed-on: http://android.intel.com:8080/134621
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>