libcamera2: [ARCH] cleanup ICameraHwInterfaces
BZ: 143142

In prior effort to align Camera HAL with abstract control APIs defined
in AtomISP refactory work the solution was to move all the public
functions of AtomISP into the newly defined abstract interfaces in
ICameraHwInterfaces.h. IHwIspControl became a trashcan for all the
functions that were not defined in other hardware controls (Sensor,
Lens or Flash). This was not the intention with APIs abstraction
designed with AtomISP refactory work and it has lead to situation
where custom additions always have impact to interfaces that aim to
be in common ground.

This change drops all the functions that do not belong to
IHwIspControl. All overloads in AtomISP are defined protected in order
to enforce using the control APIs for their purpose. The remaining
public functions in AtomISP are mainly for instantiating the V4L2
backend, controlling the streams, modes, events, etc and re-mapping
some value ranges for example. This area remains not implemented
according to target designs in AtomISP refactory and without this
change becomes more challenging.

Mainly ControlThread is impacted by this cleanup. It will need the
access to AtomISP as was before. It also uses direct IHwIspControl
which is now made explicit by using accessors in HWControlsGroup.

This change also moves the makernote file dump logic from AtomAIQ
class to CameraDump class and aims to clean the API usage into minimum
in order to allow dropping the file-injection functions from the
control interfaces.

TODO: Remove the get*Size() and getRecordingFramerate()
functions. Left in IHWIspControl to maintain patch compilation.

Change-Id: I75979267a302b926c999ca6c2d17f8cfaee5dff9
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>