libcamera2: 4 battery states for camera flash
BZ: 143408

Camera flash will be disable if power is too low. We support only 2
states previously. 0 or 1. It indicates to disable flash or not. Now the
states will be increased to 4:

NORMAL (0) – Flash ON with 1.8A
 Battery Capacity > 15% AND  Temperature > 15C
WARNING(1) – Flash ON with 1A
 Battery Capacity Less than 50% AND temperature between 15C and 5C
ALERT(2) - – Flash ON with <1A/Unused
CRITICAL(3) – Flash OFF
 Battery Capacity less than 15% OR Battery Temperature Below 5C

Camera flash will be controlled more precisely

Change-Id: Id02d30e57c65b3ad06f089b64bc5aa900e3b3f8d
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>
Reviewed-on: http://android.intel.com:8080/136717
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>