libcamera2: reduce yuv buffers number for still capture at css2.0
BZ: 138369

Issue:
Because the css2.0 requests the raw ring buffers number should >= 5,
so at the previous modification(http://android.intel.com:8080/#/c/126553/), we changed the cfg.numCaptures to 2 if the css is 2.0
But actually, it increased the yuv buffers number from 2 to 3 at the same time.

Solution:
change back the yuv buffers from 3 to 2.
add more checking in the function configureContinuousRingBuffer to meet the minimum 5 raw ring buffers request.

Depends on:
None

Change-Id: Ieabd1fb6bf3f915cbb347735a133a1b1ae5236fa
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/137181
Reviewed-by: Zheng, Hongjiang A <hongjiang.a.zheng@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Zhu, XiaofeiX <xiaofeix.zhu@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>