libcamera2: Use graphic recording buffer
BZ: 102512

This patch is going to change the recording buffer allocation place from
MemoryHeap to Graphic. Graphic memory is convenient for both software
and hardware encoder to access. It keeps the old way that share virtual
address to IntelMetaDatabuffer and would be beneficial to the encoder
that want to use Graphic buffer directly in future. Non-cacheable
gralloc buffers also would save power.

Change-Id: Ida635e964cf8d85928f76c1e50e7869fca8d677a
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/137203
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>