[PORT FROM R43S] libcamera2: adjust face metadata callback sending frequency
BZ: 145124 139069

The face metadata was sent towards the application for every preview
frame, at best at 33ms interval (30FPS). This creates some calculation
overhead at the application that draws the indicators and places them
on the UI layout.

This patch enables adjusting the  face callback frequency via the
camera_profiles.xml mechanism. The frequency can be dropped to a
fraction of preview framerate, e.g. 1/3, to allow some relief on
the application-side processing.

Change-Id: I26c3db76bef3f0fe192c5bd17caee8fc63750df0
Orig-Change-Id: I341b69707735b4071d8ed3ecb87fc2309a4f55b8
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/139731
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>