libcamera2: Differentiate burst and braketing YUV buffer number
BZ: 147156

AE or AF bracketing capture will hang sometimes due to burst capture
optimization. The max number of YUV buffer is not enough. BracketManger
has a independent thread to grab YUV data in full speed. The lower the
snapshot size, the faster the capture speed is. Sometimes it's too late
to return buffer to ISP for reuse. So bracketing needs more buffers.

Burst capture is operated in ControlThread. It always has enough time
slice to return YUV buffer. Usually 4 buffers will be enough for it. It's
to say burst capture uses same number of snapshot buffers with ZSL.

This patch will make a difference between bracketing and burst. To save
more memory for burst and to fix buffer exhausted issue of bracketing.

Change-Id: I3ff3e67c9d85817454915777bc770ab4fb7429d7
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>