libcamera2: Use general calculation for IMX132 FOV issue workaround.
BZ: 147433

The previous workaround solution for IMX132 FOV issue is only for 480P case.
This modification is to use general calculation to set proper preview sizes
for all resolutions.
The calculation uses fixed height 1080 which is full FOV size for IMX132.
If DVS is enabled, 20% envelop is cut off for preview size.

Change-Id: I3fd557d61c9ea36891d159cd2100862ed753dbc6
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>