libcamera2: set the default snapshot size to 1080p
BZ: 140905

Root cause:
Barcodes scanner doesn't care about the size of snapshot,
it use the default value that hal set.
And it will set the preview size which is matched with
screen size.
Our sensor can only output one ratio, so to choose
a ratio to match the screen is the best way to avoid stretch.

Change-Id: I3e4361a30ac5e30b7757c5940f7fbc54ed4f6a15
Signed-off-by: Yu CUI <yux.cui@intel.com>