libcamera2:[REVERTME]: selecting horizontal and vertical dynamically.
BZ: 147077

Issue:
FOV calculation test case in CtsVerifier.apk tests fail on sensor imx132.

Solution:
Get two groups of view angle from sensor vendor. Since cropping is necessary
when changing aspect ratio to 4:3, horizontal view angle is changed with
that.So the view angle should be changed along with changing aspect ratio of
image's resolution.
This is a workaround way, and It can be revert when a new feature about
this issue is uploaded in the future.

Change-Id: I42b839661db6cbdb80341097ff91e025e2950325
Signed-off-by: Li Zong <zongx.li@intel.com>