libcamera2: modify postview buffer allocation
BZ: 151967

Earlier the postview buffers were (re-)allocated in AtomISP.
This patch changes the allcation to happen in PictureThread
and these externally allocated postview buffers are set to AtomISP,
in the same manner as we do with the snapshot buffers.

This makes the postview buffer allocation and recycling scheme more uniform
with the snapshot buffer handling.

Change-Id: I5f491393865d80d265381a40293f743a758d44df
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>