libcamera2: independent recording device video3
BZ: 152961

The intention of this patch is to support full resolution SDV. We use
main device for video recording before except a special case which need
to swap it for preview device. Now for SDV, capture and video pipeline
will run at the same time, so we need a dedicated device for recording
only. It depends on driver
change:Ib086620d7730eeec512be9fcd2bf0245e9d03e18

Change-Id: I1e57c95dbbca7e6651dd36a68f19b70b50cb0ad6
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>