libcamera2: make SensorHW as common source for frame synchronization
BZ: 150082

This change moves implementation of IObserverSubject from nested class
AtomISP::FrameSyncSource into SensorHW class. In addition to sensor
related hardware controls this move makes the same abstraction
responsible of providing frame synchronization features in common as
well.

For backwards compatibility with current design of BracketManager
a new blocking call IHWSensorControl::waitForFrameSync() is provided.
This is a shortcut to combine the deprecated way that bracketing used
through AtomISP with generic use of AtomISP::FrameSyncSource observable.

Both bracketing and the remaining uses of frame synchronization events
in AAAThread and SensorSyncManager are to be addressed further in order
to encapsulate all the platform specifics under the SensorHW abstraction.

Change-Id: I8ba4abbe663727742473701467abd9072fad724f
Signed-off-by: Teemu Tuominen <teemu.tuominen@intel.com>