libcamera2: introduce new exposure synchronization logic
BZ: 150082

This change moves the old SensorSyncManager role into SensorHW
class. The managing of exposure history is changed in order to
maintain the exposure history generally for everything that requires
either timestamps or synchronization with frames mapped with certain
exposure parameters.

Such logic is fragile to scheduling and varying cpu-time available,
but is mandatory to workaround timing issues with CSS not able to
provide this information. By adding the logic in SensorHW this change
provides an opportunity to remove related workarounds from above layers.

Initially the tracked active exposure estimate is used to delay the
original SOF frame synchronization event with currently active VBI
interval in case of CSS 2.0.

Change-Id: I8db9d68de3aded27cd6cf982540876af62fd7afc
Signed-off-by: Teemu Tuominen <teemu.tuominen@intel.com>