libcamera2: add initial skip in bracketing with CSS2.0
BZ: 150082

IHWSensorControl::waitForFrameSync() replaces the old function of
AtomISP that was directly blocking the SOF event. In CSS2.0 this
frame synchronization is awaken at valid time, but it is delayed
from event that is closer to EOF. This means that initial applying
from bracketing will never reach the first frame interval and
one additional initial skip is needed.

Change-Id: I94b0288e7aa703c656748f05b1fa459eeaa7c629
Signed-off-by: Teemu Tuominen <teemu.tuominen@intel.com>