libcamera2: Video zoom failed in 1080P recording of front camera
BZ: 157074

Rootcause:
When start 1080P recording in front camera with DVS on, the sensor can't output
the frames with 20% envelope for DVS, so cause the DVS configure failed with DVS
on, then video zoom failed.

Solution:
Check if driver is enabling DVS by the envelope from ISP, if yes, configure DVS lib
with DVS on, otherwise, configure DVS lib with DVS off.

Change-Id: Idae577dfb236837565451871a242a18c8dd9d278
Signed-off-by: hongjian <hongjiang.a.zheng@intel.com>