libcamera2: call IOC for setting DVS false as well as true
BZ: 160340

When DVS is enabled, ISP will be configured with DVS ON at the
beginning. Then it should be configured to OFF when user requires
1080P@60fps because ISP has limitation that can't support both at the
same time.But there is a problematic code in AtomISP::setDVS. It will
return directly without calling IOC to disable DVS in ISP in this case.

This patch is going to fix the incorrect logic.

Change-Id: I3a75bd13d7c2ac80c83d4251e10a15dd402f8cbf
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>