libcamera2: use SensorHW exposure synchronization with bracketing
BZ: 150082

This change is loosing the old blocking call from BracketManager
and changing the bracketing design so that SensorHW exposure
synchronization logic is used for applying the sensor parameters
in parallel with BracketManager thread. This move makes the
responcibility of parameter syncronization into one place.

Change also introduces ability to pre-queue parameters in before
hand. This feature is not untilized in bracketing yet in order to
keep changes minimal.

Change-Id: I47394e125deae577fb97e1ef574530dc6e9e57d0
Signed-off-by: Teemu Tuominen <teemu.tuominen@intel.com>