libcamera2: PictureThread returns buffer in flushBuffers
BZ: 165270

The issue is sometimes snapshot buffers get lost in doing buffer flush
when user request to stop preview during capture. PictureThread
"flushBuffers" was called right after "encode" and the encoding handler
still has not started. The asynchronous encoding message was pulled
out but the snapshot buffer and postview buffer were not processed in
this case. So they are lost.
To solve this bug we need to give the buffers back to ControlThread
through pictureDone callback.

Change-Id: Ib588403d9550d42d6e80ef04eb3d9534f8e4cbca
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>