libcamera2: implement HAL ZSL by Gen7
BZ: 173316

The original camera HAL ZSL for soc sensor was implemented depends on IMG
GPU. But in BYT, we are using Gen7.
Now using Gen7 based on video vpp to implement a VAScaler for HAL ZSL HW
scaler. and it supports not only scaling but also color convert. EX. We
got YUYV format data form ISP, scale to proper size for preview and do
scaling and color convertion to NV12 for snapshot in a single call
process.
Now, it works on BYT in case that fw needs a update patch which implement
a VFPP disable mode named LOWLAT.

Change-Id: Ia36ab1829c1b21e07c3a0ee41d706421f5c7210f
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>
Signed-off-by: Yang Liang <liang.l.yang@intel.com>