libcamera2: move default template parameters to template declaration.
BZ: 171484

gcc-4.9 gives a whole lot more warnings than 4.7, so the build is broken and
we need to make proper adjustments. Particularly, gcc-4.9 starts to warn
the user when the default parameter is defined in the definition of the method,
which could be misleading to the user. Warnings are fixed by moving default
parameters to class declaration.

Change-Id: I26471afb2404f21306dd20c9462eb1e5d5241f5e
Signed-off-by: Alexander Ivchenko <alexander.ivchenko@intel.com>