libcamera2: video refactoring
BZ:171802

We do most of the recording works in ControlThread currently. It works
but seems not good from the architecture perspective. It's all agreed
that some of the work should be moved to VideoThread.

This work becomes necessary for SDV. Because we dequeue recording buffer
in ControlThread, get video snapshot buffer also in
ControlThread.Snapshot buffer capture will block for unacceptable long
time for video recording. You will see the recorded video is not
continuous in the time when you was doing SDV.

This patch is going to move video work to VideoThread. ControlThread
just does the controling work and VideoThread needs to do what it should
do.

Change-Id: I06ee6a861a3564b458d9c38032ecd25c24374818
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>