libcamera2: remove unnecessary condition in post capture done handler
BZ: 173447

handleMessagePostCaptureProcessingDone has an error logic that
introduces this bug. It judges the condition if capture sub state is IDLE
or not. If not IDLE, posts the message for next loop. We should
acknowledge that when ULL is in prcessing, the capture sub state should
never be IDLE because buffer are all in used. We didn't meet the bug in
normal ULL capture because app will call startPreview during post
processing. The capture sub state will be forcibly changed to IDLE. But
MMS is different. It will not call startPreview anymore.

This patch is going to remove the unnecessary condition processing.
Camera capture sub state is not reliable to reflect the real capture
state when multiple capture requests being processed in parallel.

Change-Id: I6dab4c8feb3dd1b69eabda2e46832bcd63dfdc62
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>