libcamera2: enable ZSL and SDV for SoC sensor with the YUVPP pipe
BZ: 183135

The ISP fw/css has added one new pipe which is called YUVPP pipe.
The YUVPP pipe is used for the SoC sensor or external ISP.
With the YUVPP pipe support, the ISP could support the SoC sensor and external ISP better.
With the YUVPP pipe, the ISP has the below capabilities:
1. the ISP could do color convert. the capability is the same as the supporting for the Raw sensor.
2. the ISP could do YUV downscaling. we could not use GPU to do downscaling.
3. the ISP could output multiple streams. for example, 3 streams for ZSL. 4 stream for SDV.

But for our current implementation for ZSL, just use one biggest stream output from the ISP.
The one stream ZSL's shortcoming is that it needs GPU to do downscaling for other streams.
If we use the YUVPP pipe, we could get all the 3 streams from driver for ZSL.

And the one stream ZSL support is not deleted.
To use one stream or multiple streams could be configured in the camera_profiles.xml
The default has been set to use one stream output.
If one platform needs multiple streams, this could be configured.

For the SDV for SoC sensor, it's similar with the ZSL. But the SDV requests four streams output.

Because there is no raw ring buffers in the driver for SoC sensor,
so for ZSL and SDV for SoC sensor, buffer queues are needed for snapshot and postview.

TODO: currently the ZSL is using four streams because of css/fw bugs.
the recording stream will be deleted for the SoC sensor's ZSL in the future.

Change-Id: I031445a31ac7ac372a31028cc72b2f5e62113a82
Signed-off-by: Yang Liang <liang.l.yang@intel.com>