libcamera2: fixes RAW bayer capture
BZ: 178189

The RAW bayer capture got broken when SensorHW classes got added.
The reason was that the same member variable was left to AtomISP
class as well, and the RAW format was not queried from the SensorHW
class, but we relied on the information in AtomISP class.

The end result was the RAW Bayer capture got stalled as we ended up
putting the postview frame back to ISP. Postview has not been supported
in RAW Bayer capture since CSS v1.5.

Change-Id: Ie4981ecce17e20c6076c6bd09a4181b041a2a5b0
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>