libcamera2: ULL image registration set to be done internally by default
BZ: 182538

There is a field in ULL configuration structure that provides information
if image registration is done externally or internally (in ia_cp_ull_compose()).
If this field is NULL, registration will be done internally.
Default value of the field is set to be NULL.

This patch depends on ICG patch
https://android.intel.com/#/c/177083/

Change-Id: I88a8bad6e08238cdae71e887fbcfa2f8a79b9b53
Signed-off-by: Dusan Zoric <dusanx.zoric@intel.com>