libcamera2: code clean up for file injection
BZ: 180498

Issue:
Currently, the file injection is not a common feature to use.
The file injection will be used just when 3A team engineers do the tunning for one new sensor.
This is one feature that just the developers use.
So we should disable it by default.

Solution:
add one macro definition, set it to 0 by default.
set it to 1 when the someone to use it.

Change-Id: I080c6f533fdf091d8625d361a9ecaeca9a2bc094
Signed-off-by: Yang Liang <liang.l.yang@intel.com>