libcamera2: add disable 3A function
BZ: 188913

Issue:
For the sensor scalability to test some raw sensors,
it reqeusts that our camera HAl code could be run without CPF file.
it means that we need to disable 3A for these test case.

Solution:
add one tag to disable 3A in the camera_profiles.xml
when it's set to true, disable the 3A to run.
by default, it's set to false.

Change-Id: I7fdef6a7bb44574a9ab0e87677275c552afb9c0c
Signed-off-by: Yang Liang <liang.l.yang@intel.com>