libcamera2: ULL with SRES feature
BZ: 190060

Patch should enable ULL with SRES feature.
When ULL is ON, we are setting SRES mode in driver
through new IOCTL(ATOMISP_IOC_S_ENABLE_DZ_CAPT_PIPE).
In this mode, IPU will zoom preview frames, but not capture frames.
We are sending capture snapshots to ULL alongside with zoom factor.
For reference first frame is copied and zoomed with bicubic upscale.

Change-Id: I39ac3de4a680f613b3f86baf6e49df169f5f0620
Signed-off-by: Kostic Dusan <dusan.kostic@intel.com>