libcamera2: HAL Video Stabilization support
BZ: 187685

0-copy hal video stabilization support.

This allows to integrate a video stabilization algorithm in the HAL.
The camera is configured for one output only. Preview and video
share the same output buffers from the video stabilization in 0-copy
mode of operation.

Currently the "videostabilization" is just a memcpy, the customer
needs to replace that with their implementation.

Camera buffers are not using yet a bigger than preview size
resolution with the envelope required by the VS algorithm.

In order to enable HAL VS, the camera_profiles.xml needs to have
tag "useHALVideoStabilization" as true, and preview + video
resolutions must match, and video stabilization must be enabled
in the application.

Change-Id: I5de0abdbacd04d61a9a59463c3c2546289c082ca
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>