libcamera2: Fix GPUWarper bug that happens when frames are processed in tiles
BZ: 186430

There is a limit on size of input texture that can be used in OpenGL
(GL_MAX_TEXTURE_SIZE). Value of this parametes is 4096 on SGX GPU, and 8192 on
RGX GPU.

When one of the picture dimensions is larger than GL_MAX_TEXTURE_SIZE,
picture should be divided into matrix of tiles in order to be processed.

To obtain warped output tile (depending on the motion matrix used), besides
input tile pixels, some additional (border) pixels are needed. If warping
operation is performed in-place, i.e. if warped tile is stored in the input
tile buffer, currently warped tile can not be stored before its adjacent tiles
to be processed are read. In the previous implementation that was not taken
into account, so that, in some scenarios, incorrect results could be obtained.

This was fixed by allowing division of input frames only in one (horizontal)
demension (so that each tile has up to 2 adjacent tiles). Also, next tile to
be processed is now uploaded before previously warped tile is stored.

Change-Id: I8ecc851627d37d79a9844049cfb96b206ec64f5b
Signed-off-by: Dusan Zoric <dusanx.zoric@intel.com>