libcamera2: envelope sizes for HAL video stabilization
BZ: 187685

This patch adds an envelope size to the HAL video stabilization
feature. The driver is configured with the envelope size, which is
bigger than preview, and then the video stabilization class takes
care of cropping the input to the preview buffers.

Customer should integrate their own video stabilization into the
class HALVideoStabilization, which currently only has a simple
cropping algorithm which works for NV12 and NV21.

This patch can be tried currently with the reference mofd device
only, and it has the limitation that there is no working envelope
resolution for 1080p. Thus 720p (1280x720) is the supported video
resolution in the reference device. In practice camera_profiles.xml
must now include the preview resolution "1280x720" for this to work.
Also hal videostabilization must be enabled in camera_profiles.xml,
as previously.

This also restores the preview callbacks for HAL VS and adds simple
crop functionality for NV12/NV21.

Change-Id: Icc69819674b6e6c5706b3828f0e6dfc59fa0ffe9
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>