libcamera2: add the support for extended camera
BZ: 189975

Issue:
for the Kevlar project which is based on the MOFD, it has totally 3 sensors.
the back sensor is ov8858 (raw)
the front sensors are imx132 which is raw sensor
and the other one is ov680 which is soc.
and the ov680 is one special sensor which is used to measure the depth.
the first step for Kevlar project is to let all three camera could be work for doing the tuning seperatelly.
so this is one MUST work.
but the Google architecture just support 2 sensors, 0 is back, 1 is front.
so we need to add one extended camera.

Solution:
before the calling of the camera HAL,
we should let the camera HAL to know that if the extended camera is running.
We couldn't add one Java API, because it breaks the Google CDD.
so we add one debug interface to do it.
in the camera_profiles.xml file, we add two new tags:sensorName, extendedCamera
in the PlatformData, there is one mapping to find the real setting
which is stored in the PlatformData.mCameras

Change-Id: I5a5ddef49894f87c686b713232656bd9bedcb6b5
Signed-off-by: Yang Liang <liang.l.yang@intel.com>