[PORT FROM MAIN]libcamera2: move the mPreviewFourcc from Common section to the Sensor section
BZ: 208935

Issue:
the preview format setting is set by platform right now,
it means that the preview format will be set to the same for front camera and back camera.
but the ISP output maybe different by sensors.

Solution:
we need to move the mPreviewFourcc from Common section to the Sensor section.

Change-Id: I4618cfda4f1489c6f1420ad08260ca566c6bb632
Signed-off-by: Yang Liang <liang.l.yang@intel.com>