[PORT FROM R44C] libcamera2: fix the problem in the disable3A patch
BZ: 202259

Issue:
in the patch 178016 which is used to disable 3A function,
there are some issues.
this patch is used to fix the issues in that patch.

Solution:
1. when the RAW sensor is used and we need to disable 3A,
we will create an AtomSoc3A object instead of AtomAIQ object.
this will bypass the intel 3A function.
2. in the AAAThread, when we need to disable 3A,
it doesn't need to handle the real message handling functions.
3. the frame sync event is unnecessary when the disable 3A is set

Change-Id: I25eb1e9fba304c1af33689dacfcaa9db013b73d3
Signed-off-by: Xia ShouYong <shouyongx.xia@intel.com>
Signed-off-by: Yang Liang <liang.l.yang@intel.com>