libcamera2: Implement offline bracket by locking raw buffer
BZ:192193

It's very hard to get a specified snapshot buffer we need with current
CSS APIs which is the biggest problem in current offline bracketing.
e.g. we set the exposure bias -2/0/2 from the SOF of image with exposure
id 3. We know it will be applied from the image with exposure id 5. But
it's not accurate to use offset of continuous capture config to grab
buffer 5.

Now we can use css raw buffer lock mechanism to lock the raw buffer we
want and capture the buffer with exposure id. It helps to get the right
buffer we need.

Currently it is used to implement offline mode AE bracketing only

Depends on kernel changes:
1. Id: Id1fd403f532a9e2c56082e2c36a8ca81f7fb3a2c
2. Id: I989ae8a0435b42818053e7d54d782d48c6d01870
3. Id: I9ccf2e00eca215a433a23ee883d42f74319e1b09

Change-Id: I6a8e78dc3671d1d8fb6f71e6214ea2c66c848f39
Signed-off-by: Xie,Hongbiao <hongbiao.xie@intel.com>