libcamera2: Clear snapshot and Postview buffers when stop capture
BZ: 189797

Sometimes if ISP timeout happen, HAL will try to recover it.
It will call stopCapture to clear mUsingClientSnapshotBuffers
and set mConfig.num_snapshot to 0. When initialize ISP in recovering process,
mConfig.num_snapshot will be set to 1. If exit camera now,
it will cause fatal crash in freeSnapshotBuffers.
The root cause is that we free mSnapshotBuffer twice in this case.

Solution:
Clear snapshot and Postview buffers when clear related flags.

Change-Id: I55553db027f0f0194c1ca5fcd8a0438b545be915
Signed-off-by: Zhang Qingwu <qingwu.zhang@intel.com>