[PORT FROM R44C] libcamera2: send only the YUV422 data during ext-isp HDR & LLS
BZ: 196103

This patch changes the ext-isp callback sending so that only the
YUV422 data from within the buffer is sent to the application.

This is achieved by taking the buffer from inside the
CameraHeapMemory object and by wrapping it into an internal
class which is also derived from MemoryBase, and by overriding
the size and offset values then so that only a part of the
heap allocated memory is sent during the callback. After the
callback the original MemoryBase buffer is restored to the
CameraHeapMemory object.

Change-Id: Ie4b2e5383bba201a98067cfe3530871db3c0959c
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>