[PORT FROM R44C] libcamera2: initialize mHALVideoStabilization
BZ: 196104

Adding initialization of mHALVideoStabilization to prevent
camera from crashing at startup. HAL VS only works for
1280x720 and 1920x1080 currently, and now it was randomly
turning on for the still mode 640x480 preview. The envelope
resolution is not in supported list for 640x480 and the
resolution validation then failed, causing setParameters to
fail and the application to exit.

Change-Id: I7fe8ff61b59eab00c85e35102af7bc414492dc92
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>