[PORT FROM R44C] libcamera2: add the support for the low-power vision sensing mode
BZ: 201642

it's the feature for the front camera.
the feature's request is as below.
1. the front sensor output 184*104 raw8 data with 30fps, but 256*104 will be output from the ISP.
there are (256-184)bits padding at the rear of every line.
2. before the feature starts to work, the parameter "intelligent-mode" will be set to camera HAL.
3. then the camera hal knows the mode is low-power vision sensing mode.
4. in the preview callback handling, the padding will be removed when the memory copy is in process.
5. when the data is ready, it will be return with the callback CAMERA_MSG_PREVIEW_FRAME, and the preview format is one private defined.
so left it to do in the future.
6. the windowless mode will be used.

when the feature is enabled, the 3A must be disabled, because the data from ISP is raw8, the 3A could handle the raw data.

in the patch, the supported value of intelligent-mode will be read from camera_profiles.xml
for the front camera, it will be set to "true,flase".
then the supported value will be got by the application.
if the application wants to enable the feature, it will set the special preview size to camera HAL.
and the "true" of "intelligent-mode" will be set to camera HAl too.
no surface will be set to camera HAL.
the data from the driver will be sent to application by the CAMERA_MSG_PREVIEW_FRAME.

Change-Id: Ibe17fdddf58d91228e81d675bb0ca2dd06e4dcc0
Signed-off-by: Yang Liang <liang.l.yang@intel.com>