[PORT FROM R44C] libcamera2: provide LLS need to application via face callback
BZ: 202256

This adds the LLS need to the face callback. The camera service
layer needs a change (customer will do the change) for this to
work all the way to java; operation will be as previously
without such a change. Adding the new field to the end of the
face struct does not cause issues.

LLS cmd id's are updated also by this patch.

Change-Id: Ie39c92f930c58a01653eb87a66c1c88dbb16c6ea
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>