libcamera2: Fix camera profiles reading for depth camera
BZ: 192393

Depth camera extension was broken due to conflicts with multi-sensor
support. This patch fixes the conflict without interface changes.

The original new tags sensorName and extendedCamera are removed,
and replaced with attributes. In camera_profiles.xml thus the depth
sensor should be denoted with:

<Profiles CameraId="1" name="ov680" extension="depth">

The sensor name is optional.

Change-Id: Ic6732926c544a5545fe6ee60fe31b0e0966ace55
Signed-off-by: Jarno Jakonen <jarno.jakonen@intel.com>
Signed-off-by: Zhang Qingwu <qingwu.zhang@intel.com>