[PORT FROM R44C] libcamera2: fixes starting subsequent AF sequences with M10MO
BZ: 202244

Previously we did not wait M10MO to report new AF sequence start
in the NV12 metadata. Instead we were using the old AF sequence information
that is still present in the metadata at the start of the new AF run,
which caused premature end of the new AF sequence and reporting
the wrong result to application.

This patch addresses the issue by adding a flag to wait for the
metadata to report that a new AF sequence has started on M10MO.

Change-Id: I9648c47be06a614e5c4aad74a0a3cc1ccbc53688
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>